add another key reference
the paper takes Austronesian language family as an example and goes through advantages and disadvantages of using Bayesian tree-building methods. It is a nice and easy read for people unfamiliar with computational methods in historical linguistics, kind of introduction to Bayesian statistics for tree-building. There is also a small section on ancestral state reconstruction, which is a nice bonus. It would be good if students could read or at least look through it in advance.